Fixes after review.
