+++ apps/portals-example/.env.local.example
@@ -7,7 +7,6 @@ DIAL_API_KEY=your-api-key
 DEFAULT_MODEL=gpt-4-turbo # Default model ID to use when none specified
 
 SDMX_API_URL=https://your-sdmx-api-endpoint.com
-SDMX_PLUS_URL=https://your-sdmx-plus-endpoint.com
 SDMX_AUTH_KEY=your-key
 
 # Authentication (if enabled)
