+++ apps/portals-example/src/app/api/api.ts
@@ -12,9 +12,6 @@ export const sdmxApiClient = new SdmxApiClient({
   apiUrl: process.env.SDMX_API_URL || '',
   constrainsApiUrl: process.env.CONSTRAINS_SDMX_API_URL || '',
   sdmxProxyUrl: process.env.SDMX_PROXY_URL || '',
-  sdmxPlusUrl: process.env.SDMX_PLUS_URL || '',
-  codelistUrl: process.env.SDMX_CODELIST_URL,
-  hierarchyUrl: process.env.SDMX_HIERARCHY_URL,
   apiKey: process.env.SDMX_AUTH_KEY,
 });
