+++ libs/sdmx-toolkit/src/utils/hierarchy-utils.ts
@@ -3,7 +3,6 @@ import {
   CodelistData,
   Hierarchy,
   HierarchicalCode,
-  TreeNode,
 } from '../models/structural-metadata/hierarchy';
 import { DataConstraints } from '../models/structural-metadata/constraints';
 import { StructuralData } from '../models/structural-metadata/structural-metadata';
@@ -52,7 +51,11 @@ export function urnMatchesIgnoreVersion(urnA: string, urnB: string): boolean {
 }
 
 function getCodelistShortUrn(codelist: CodelistData): string {
-  return generateShortUrn(codelist.id, codelist.version, codelist.agencyID);
+  return generateShortUrn(
+    codelist.id,
+    codelist.version ?? '',
+    codelist.agencyID ?? '',
+  );
 }
 
 function collectCodes(
@@ -122,94 +125,4 @@ export function getHierarchyAvailableCodes(
 
   const validIds = new Set(keyValues.selectionValues.map((v) => v.memberValue));
   return codes.filter((c) => validIds.has(c.id));
-}
-
-function buildNodes(
-  codeListMap?: Record<string, CodelistItemBase[]>,
-  availableCodes?: string[],
-  dimensionCodeListUrn?: string,
-  codes?: HierarchicalCode[],
-  parentUrn?: string,
-): TreeNode<HierarchicalCode>[] {
-  return (
-    codes
-      ?.map((code) => {
-        const { childId, agency, id, version } = getChildParsedUrn(code.code);
-        const codelistShortUrn = generateShortUrn(id, version, agency);
-
-        // Resolve the actual codelist key ignoring version mismatches between
-        // the hierarchy reference and the codelist that was returned.
-        const resolvedUrn =
-          Object.keys(codeListMap ?? {}).find((key) =>
-            urnMatchesIgnoreVersion(key, codelistShortUrn),
-          ) ?? codelistShortUrn;
-        const items = codeListMap?.[resolvedUrn];
-        const item = items?.find((c) => c.id === childId);
-        const displayName = item?.name || item?.description || childId;
-
-        const isFromDimensionCodelist = dimensionCodeListUrn
-          ? urnMatchesIgnoreVersion(codelistShortUrn, dimensionCodeListUrn)
-          : true;
-
-        // Whether this node was originally a structural/grouping node
-        // (had children in the source data, regardless of pruning).
-        const isStructuralNode =
-          !isFromDimensionCodelist && (code.hierarchicalCodes?.length ?? 0) > 0;
-
-        const children = buildNodes(
-          codeListMap,
-          availableCodes,
-          dimensionCodeListUrn,
-          code.hierarchicalCodes,
-          code.code,
-        );
-
-        const isLeaf = children.length === 0;
-
-        // Prune structural grouping nodes whose children were all pruned.
-        if (isStructuralNode && isLeaf) return null;
-
-        // Prune dimension-codelist leaf nodes that are outside the
-        // constraint-available set.
-        if (
-          isFromDimensionCodelist &&
-          isLeaf &&
-          availableCodes &&
-          !availableCodes.includes(childId ?? '')
-        )
-          return null;
-
-        return {
-          id: code.code,
-          name: displayName,
-          children,
-          isExpanded: true,
-          // A non-dimension node is non-selectable only when it has no enabled
-          // (dimension) descendants; if it wraps selectable children it acts
-          // as a group checkbox and must itself be enabled.
-          disabled:
-            !isFromDimensionCodelist && children.every((c) => c.disabled),
-          parent: parentUrn,
-          metadata: { ...code, hierarchicalCodes: undefined },
-        } as TreeNode<HierarchicalCode>;
-      })
-      .filter((node): node is TreeNode<HierarchicalCode> => node !== null) ?? []
-  );
-}
-
-export function getTreeNodesFromHierarchies(
-  hierarchy?: Hierarchy,
-  codeListMap?: Record<string, CodelistItemBase[]>,
-  availableCodes?: string[],
-  dimensionCodeListUrn?: string,
-): TreeNode<HierarchicalCode>[] {
-  if (!availableCodes || availableCodes.length === 0) return [];
-
-  return buildNodes(
-    codeListMap,
-    availableCodes,
-    dimensionCodeListUrn,
-    hierarchy?.hierarchicalCodes,
-    undefined,
-  );
 }
